ER-880 - Sharing Skills view model behaviour
