+++ src/Provider/Provider.Web/ViewModels/Part2/Skills/SkillsEditModel.cs
@@ -1,9 +1,10 @@
 ﻿using System.Collections.Generic;
 using Esfa.Recruit.Provider.Web.RouteModel;
+using Esfa.Recruit.Shared.Web.ViewModels.Skills;
 
 namespace Esfa.Recruit.Provider.Web.ViewModels.Part2.Skills
 {
-    public class SkillsEditModel : VacancyRouteModel
+    public class SkillsEditModel : VacancyRouteModel, ISkillsEditModel
     {
         public List<string> Skills { get; set; }
 
@@ -13,8 +14,6 @@ namespace Esfa.Recruit.Provider.Web.ViewModels.Part2.Skills
 
         public string AddCustomSkillName { get; set; }
 
-        public bool IsAddingCustomSkill => !string.IsNullOrEmpty(AddCustomSkillAction);
-
-        public bool IsRemovingCustomSkill => !string.IsNullOrWhiteSpace(RemoveCustomSkill);
+        public string AddCustomSkill => !string.IsNullOrEmpty(AddCustomSkillAction) ? AddCustomSkillName : null;
     }
 }
