+++ src/Shared/Recruit.Shared.Web/ViewModels/Skills/ISkillsEditModel.cs
@@ -1,0 +1,11 @@
+﻿using System.Collections.Generic;
+
+namespace Esfa.Recruit.Shared.Web.ViewModels.Skills
+{
+    public interface ISkillsEditModel
+    {
+        List<string> Skills { get; set; }
+        string RemoveCustomSkill { get; }
+        string AddCustomSkill { get; }
+    }
+}
