+++ src/Shared/Recruit.Shared.Web/ViewModels/Skills/ISkillsViewModel.cs
@@ -1,0 +1,10 @@
+﻿using System.Collections.Generic;
+
+namespace Esfa.Recruit.Shared.Web.ViewModels.Skills
+{
+    public interface ISkillsViewModel
+    {
+        List<SkillViewModel> Column1Checkboxes { get; set; }
+        List<SkillViewModel> Column2Checkboxes { get; set; }
+    }
+}
