+++ src/Shared/Recruit.Shared.Web/ViewModels/Skills/SkillViewModel.cs
@@ -1,0 +1,9 @@
+﻿namespace Esfa.Recruit.Shared.Web.ViewModels.Skills
+{
+    public class SkillViewModel
+    {
+        public string Name { get; set; }
+        public bool Selected { get; set; }
+        public string Value { get; set; }
+    }
+}
